правки по 2 ревью

+++ src/wishlists/wishlists.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, UnauthorizedException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Wishlist } from '../entity/Wishlist';
@@ -62,20 +66,21 @@ export class WishlistsService {
       where: { id: id },
       relations: ['items', 'owner'],
     });
-    if (user.id === wishlist.owner.id) {
-      const items = await this.wishesService.findMany(
-        updateWishlistDto.itemsId,
-      );
+    if (user.id === wishlist.owner.id && updateWishlistDto.itemsId) {
       // Удаляем желания не из списка
       // Проходим по всем желаниям (items)
       wishlist.items.map((wishItem) => {
         if (!updateWishlistDto.itemsId.includes(wishItem.id))
           wishlist.items = wishlist.items.filter((item) => item !== wishItem);
       });
+      // Добавляем в items желания из списка
+      const wishes = await this.wishesService.findMany(
+        updateWishlistDto.itemsId,
+      );
+      wishlist.items.push(...wishes);
       await this.wishlistRepository.save(wishlist);
       await this.wishlistRepository.update(id, {
         ...updateWishlistDto,
-        items,
       });
       const newList = await this.wishlistRepository.findOne({
         where: { id: id },
@@ -86,16 +91,19 @@ export class WishlistsService {
       newList.description = undefined;
       return newList;
     }
-    return UnauthorizedException;
+    throw new UnauthorizedException('Список не получается отредактировать');
   }
   async delete(id: number) {
     const newList = await this.wishlistRepository.findOne({
       where: { id: id },
       relations: ['items', 'owner'],
     });
-    newList.owner.email = undefined;
-    newList.itemsId = undefined;
-    await this.wishlistRepository.delete(id);
-    return newList;
+    if (newList) {
+      newList.owner.email = undefined;
+      newList.itemsId = undefined;
+      await this.wishlistRepository.delete(id);
+      return newList;
+    }
+    throw new NotFoundException('Список не найден');
   }
 }
